pyproject.toml
[project]
name = "aiops-alerts"
version = "0.1.0"
description = ""
authors = [
    {name = "ShreyeshSwargam98",email = "shreyesh.swargam@bootlabstech.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "ollama (>=0.5.3,<0.6.0)",
    "weaviate-client (>=3.26.7,<4.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
